requirements-jetson.txt
# Core DL stack (downgraded for JetPack 5.1.2 compatibility)
torchmetrics==1.5.2

# Media and processing
av==12.0.0
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.1.1
cycler==0.12.1
filelock==3.13.3
fonttools==4.50.0
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
idna==3.6
importlib-resources==6.4.0
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
opencv-python==4.5.4.60 # Pinned to match JetPack 5.1.2
packaging==24.0
pandas==2.0.3
pillow==10.3.0
psutil==5.9.8
py-cpuinfo==9.0.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
scipy==1.10.1
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
thop==0.1.1.post2209072238
tqdm==4.66.2
typing-extensions==4.10.0
tzdata==2024.1
ultralytics==8.0.16  # Updated for torch 2.1.0 compatibility
urllib3==2.2.1
zipp==3.20.2

# Model acceleration
accelerate==1.0.0
safetensors==0.4.5
huggingface-hub==0.25.1
timm==1.0.19  # Updated for torch 2.1.0 compatibility
lightning-utilities==0.11.7
pycocotools==2.0.7

# Misc + tracking
setproctitle==1.3.3
platformdirs==4.3.6
sentry-sdk==2.15.0
click==8.1.7
docker-pycreds==0.4.0
wandb==0.18.3

# Dev and gRPC
pytest
grpcio==1.56.2
grpcio-tools==1.56.2
protobuf==4.23.4 # Downgraded for ONNX + TensorRT compatibility
python-signal==0.0.2
rich==14.0.0
gputil==1.4.0
googleapis-common-protos==1.69.2
asciichartpy==1.5.25
scikit-learn==1.3.2

# ONNX + TensorRT inference stack
onnx==1.14.1
onnxruntime==1.16.3
# pycuda  # Install with apt: `sudo apt install python3-pycuda`
